Review tab - Move getFlavorText and refactor getFlavorValue
