+++ frontend/packages/kubevirt-plugin/src/components/create-vm-wizard/tabs/review-tab/utils.ts
@@ -2,12 +2,13 @@ import { FormFieldType } from '../../form/form-field';
 import { getBooleanReadableValue } from '../../../../utils/strings';
 import { iGetFieldValue } from '../../selectors/immutable/field';
 import { VMSettingsField } from '../../types';
-import {iGet, iGetIn, toShallowJS} from '../../../../utils/immutable';
+import { iGet, iGetIn, toShallowJS } from '../../../../utils/immutable';
 import { CUSTOM_FLAVOR } from '../../../../constants/vm';
 import { iGetRelevantTemplate } from '../../../../selectors/immutable/template/combined';
 import { VMTemplateWrapper } from '../../../../k8s/wrapper/vm/vm-template-wrapper';
 import { Map as ImmutableMap } from 'immutable';
 import { ITemplate } from '../../../../types/template';
+import { getFlavorText } from '../../../../selectors/vm/flavor-text';
 
 export const getReviewValue = (field: any, fieldType: FormFieldType) => {
   const value = iGetFieldValue(field);
@@ -19,7 +20,8 @@ export const getReviewValue = (field: any, fieldType: FormFieldType) => {
 
 export const getField = (key: VMSettingsField, vmSettings) => iGet(vmSettings, key);
 
-export const getFieldValue = (vmSettings, key: VMSettingsField) => iGetIn(vmSettings, [key, 'value']);
+export const getFieldValue = (vmSettings, key: VMSettingsField) =>
+  iGetIn(vmSettings, [key, 'value']);
 
 export const getFlavorValue = ({
   iVMSettings,
@@ -28,10 +30,10 @@ export const getFlavorValue = ({
   relevantOptions,
 }: GetFlavorValueParams) => {
   const flavor = getFieldValue(iVMSettings, VMSettingsField.FLAVOR);
-  let cpuCores, memory;
+  let cpu, memory;
 
   if (flavor === CUSTOM_FLAVOR) {
-    cpuCores = getFieldValue(iVMSettings, VMSettingsField.CPU);
+    cpu = getFieldValue(iVMSettings, VMSettingsField.CPU);
     memory = getFieldValue(iVMSettings, VMSettingsField.MEMORY);
   } else {
     const template_ = toShallowJS(
@@ -39,11 +41,11 @@ export const getFlavorValue = ({
     );
     const template = new VMTemplateWrapper(template_, true).init().clearRuntimeMetadata();
 
-    cpuCores = template.getCPU()?.cores;
+    cpu = template.getCPU();
     memory = template.getMemory();
   }
 
-  return `${flavor}: ${cpuCores} CPU, ${memory}`;
+  return getFlavorText({ cpu, memory, flavor });
 };
 
 type GetFlavorValueParams = {
